requirements.txt
auto_mix_prep==0.2.0
dictor==0.1.5
diskcache==5.3.0
lightning==1.9.5
matplotlib==3.5.3
mmcv==2.0.0
mmcv_full==1.5.3
mmdet==2.25.0
numpy==1.18.1
omegaconf==2.3.0
open3d==0.9.0.0
opencv_python==4.2.0.32
pandas==0.25.3
Pillow==9.5.0
plyfile==0.9
py_trees==0.8.3
pyephem==9.99
pygame==2.0.1
pyrootutils==1.0.4
Requests==2.31.0
rich==13.3.5
rosgraph_msgs==0.0.1
scikit_image==0.16.2
scipy==1.4.1
setuptools==65.6.3
Shapely==1.7.0
simplejson==3.19.1
six==1.14.0
tabulate==0.8.7
tf==1.0.0
timm==0.5.4
torch==1.13.1
torch_scatter==2.1.1
torchvision==0.14.1
tqdm==4.43.0
trimesh==3.21.5
ujson==5.3.0
wandb==0.8.29
xmlschema==1.0.18
